requirements.txt
python-dotenv==1.0.1
twitchio==2.8.2
